% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ConductorFailureTimes}
\alias{ConductorFailureTimes}
\title{Dataset: ConductorFailureTimes}
\format{
A numeric vector of failure times.
}
\usage{
data(ConductorFailureTimes)
}
\description{
This dataset contains failure times measured in hours from an accelerated life test with 59 conductors.
}
\details{
This dataset contains failure times (measured in hours) obtained from an accelerated life test involving 59 conductors.
The data are presented as a numeric vector.
}
\references{
Nasiri, B., et al. (2010). "Bayesian analysis of the accelerated life model with Type-II censoring."
Journal of Statistical Planning and Inference, 140(6), 1565-1572.

Schafft, H. A., et al. (1987). "Reproducibility of the accelerated test for electric cable insulation."
IEEE Transactions on Electrical Insulation, 22(5), 739-746.
}
\keyword{datasets}
