% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTLCAR.R
\name{fTLCAR}
\alias{fTLCAR}
\title{Estimate parameters for the TLCAR distribution using maximum likelihood.}
\usage{
fTLCAR(data)
}
\arguments{
\item{data}{Numeric vector of data values.}
}
\value{
Numeric vector of estimated parameters.
}
\description{
This function estimates the parameters of the TLCAR distribution while
respecting the constraints on the parameters.
}
\examples{
data(ConductorFailureTimes)
estimated_params <- fTLCAR(ConductorFailureTimes)

}
