% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{qTLCAR}
\alias{qTLCAR}
\title{Quantile function for TLCAR distribution}
\usage{
qTLCAR(p, alpha, a, b, theta, m)
}
\arguments{
\item{p}{Probability value (between 0 and 1).}

\item{alpha}{Parameter representing the distribution of the Topp-Leone component.}

\item{a}{Parameter representing the scale (a) of the Cauchy component.}

\item{b}{Parameter representing the position (b) of the Cauchy component.}

\item{theta}{Parameter representing the scale of the Rayleigh component.}

\item{m}{Additional parameter.}
}
\value{
Numeric value representing the quantile.
}
\description{
Calculate the quantile value for a given probability using the TLCAR distribution.
}
\examples{
 qTLCAR(p = 0.5, alpha = 1, a = 1, b = 0, theta = 3, m = 1)

}
