% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{rTLCAR}
\alias{rTLCAR}
\title{Generate a random sample from the TLCAR distribution}
\usage{
rTLCAR(n, alpha, a, b, theta, m)
}
\arguments{
\item{n}{Number of observations in the sample.}

\item{alpha}{Parameter representing the distribution of the Topp-Leone component.}

\item{a}{Parameter representing the scale (a) of the Cauchy component.}

\item{b}{Parameter representing the position (b) of the Cauchy component.}

\item{theta}{Parameter representing the scale of the Rayleigh component.}

\item{m}{Additional parameter.}
}
\value{
Random sample from the TLCAR distribution.
}
\description{
Generate a random sample from the TLCAR distribution using the quantile function.
}
\examples{
# Generate a random sample with 100 observations using estimated parameters
sample <- rTLCAR(n = 100, alpha = 1, a = 1, b = 0, theta = 3, m = 1)

}
