% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempvar.R
\name{temp_var}
\alias{temp_var}
\title{Temporary Variable Calculation}
\usage{
temp_var(x, theta, a, b, m)
}
\arguments{
\item{x}{Numeric vector of values at which to calculate the temporary variable.}

\item{theta}{Parameter representing the scale of the Rayleigh component.}

\item{a}{Parameter representing the scale (a) of the Cauchy component.}

\item{b}{Parameter representing the position (b) of the Cauchy component.}

\item{m}{Additional parameter.}
}
\value{
Numeric vector of calculated temporary variable values
}
\description{
This function calculates a temporary variable used in the TLCAR distribution density function.
}
