% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{Tree_diameters}
\alias{Tree_diameters}
\title{Dataset: Tree_diameters}
\format{
A numeric vector of tree diameter measurements (in cm).
}
\usage{
data(Tree_diameters)
}
\description{
This dataset contains tree diameter measurements (in cm) for Teak trees in the Agrimey sector in Benin.
}
\keyword{datasets}
